Add a BmpImageWriter
