+++ image/shared/src/main/scala/eu/joaocosta/minart/graphics/image/bmp/BmpImageLoader.scala
@@ -1,0 +1,109 @@
+package eu.joaocosta.minart.graphics.image.bmp
+
+import java.io.InputStream
+
+import scala.annotation.tailrec
+
+import eu.joaocosta.minart.graphics._
+import eu.joaocosta.minart.graphics.image._
+import eu.joaocosta.minart.graphics.image.helpers._
+
+/** Image loader for BMP files.
+  *
+  * Supports uncompressed 24/32bit Windows BMPs.
+  */
+trait BmpImageLoader[F[_]] extends ImageLoader {
+  val byteReader: ByteReader[F]
+  import byteReader._
+
+  private val loadRgbPixel: ParseState[String, Color] =
+    readBytes(3)
+      .collect(
+        { case bytes if bytes.size == 3 => Color(bytes(2), bytes(1), bytes(0)) },
+        _ => "Not enough data to read RGB pixel"
+      )
+
+  private val loadRgbaPixel: ParseState[String, Color] =
+    readBytes(4)
+      .collect(
+        { case bytes if bytes.size == 4 => Color(bytes(2), bytes(1), bytes(0)) },
+        _ => "Not enough data to read RGBA pixel"
+      )
+
+  @tailrec
+  private def loadPixels(
+      loadColor: ParseState[String, Color],
+      data: F[Int],
+      remainingPixels: Int,
+      acc: List[Color] = Nil
+  ): ParseResult[List[Color]] = {
+    if (isEmpty(data) || remainingPixels == 0) Right(data -> acc.reverse)
+    else {
+      loadColor.run(data) match {
+        case Left(error)               => Left(error)
+        case Right((remaining, color)) => loadPixels(loadColor, remaining, remainingPixels - 1, color :: acc)
+      }
+    }
+  }
+
+  def loadHeader(bytes: F[Int]): ParseResult[Header] = {
+    (for {
+      magic <- readString(2).validate(
+        BmpImageFormat.supportedFormats,
+        m => s"Unsupported format: $m. Only windows BMPs are supported"
+      )
+      size   <- readLENumber(4)
+      _      <- skipBytes(4)
+      offset <- readLENumber(4)
+      dibHeaderSize <- readLENumber(4).validate(
+        dib => dib >= 40 && dib <= 124,
+        dib => s"Unsupported DIB header size: $dib"
+      )
+      width  <- readLENumber(4)
+      height <- readLENumber(4)
+      colorPlanes <- readLENumber(2).validate(
+        _ == 1,
+        planes => s"Invalid number of color planes (must be 1): $planes"
+      )
+      bitsPerPixel <- readLENumber(2).validate(
+        Set(24, 32),
+        bpp => s"Unsupported bits per pixel (must be 24 or 32): $bpp"
+      )
+      compressionMethod <- readLENumber(4).validate(
+        c => c == 0 || c == 3 || c == 6,
+        _ => "Compression is not supported"
+      )
+      loadColorMask = compressionMethod == 3 || compressionMethod == 6
+      _         <- if (loadColorMask) skipBytes(20) else noop
+      redMask   <- if (loadColorMask) readLENumber(4) else State.pure[F[Int], Int](0x00ff0000)
+      greenMask <- if (loadColorMask) readLENumber(4) else State.pure[F[Int], Int](0x0000ff00)
+      blueMask  <- if (loadColorMask) readLENumber(4) else State.pure[F[Int], Int](0x000000ff)
+      _         <- if (loadColorMask) skipBytes(4) else noop // Skip alpha mask (or color space)
+      _ <- State.check(
+        redMask == 0x00ff0000 && greenMask == 0x0000ff00 && blueMask == 0x000000ff,
+        "Unsupported color format (must be either RGB or ARGB)"
+      )
+      header = Header(magic, size, offset, width, height, bitsPerPixel)
+      _ <- skipBytes(offset - (if (loadColorMask) 70 else 34))
+    } yield header).run(bytes)
+  }
+
+  def loadImage(is: InputStream): Either[String, RamSurface] = {
+    val bytes = fromInputStream(is)
+    loadHeader(bytes).right.flatMap { case (data, header) =>
+      val numPixels = header.width * header.height
+      val pixels = header.bitsPerPixel match {
+        case 24 =>
+          loadPixels(loadRgbPixel, data, numPixels)
+        case 32 =>
+          loadPixels(loadRgbaPixel, data, numPixels)
+        case bpp =>
+          Left(s"Invalid bits per pixel: $bpp")
+      }
+      pixels.right.flatMap { case (_, flatPixels) =>
+        if (flatPixels.size != numPixels) Left(s"Invalid number of pixels: Got ${flatPixels.size}, expected $numPixels")
+        else Right(new RamSurface(flatPixels.sliding(header.width, header.width).toSeq.reverse))
+      }
+    }
+  }
+}
