+++ image/shared/src/main/scala/eu/joaocosta/minart/graphics/image/bmp/Header.scala
@@ -1,0 +1,10 @@
+package eu.joaocosta.minart.graphics.image.bmp
+
+final case class Header(
+    magic: String,
+    size: Int,
+    offset: Int,
+    width: Int,
+    height: Int,
+    bitsPerPixel: Int
+)
